+++ src/index.ts
@@ -1,1 +1,4 @@
 export * from './k-number';
+
+export * from './constants';
+export * from './utils';