+++ src/utils/index.ts
@@ -1,0 +1,1 @@
+export const isInteger = (num: number) => num % 1 === 0;